# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Movia Immo - API de gestion immobilière
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      parameters:
        - in: query
          name: zone
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Property list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a tenant by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    put:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tenants/{id}/payments:
    get:
      operationId: getTenantPayments
      tags: [tenants]
      summary: Get payments for a specific tenant
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  # Payments
...
components:
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [appartement, villa, bureau, local_commercial, studio, riad]
        zone:
          type: string
          enum: [Malabata, Cap_Spartel, Medina, Asilah, Tetouan, Martil, Fnideq, Centre_Ville, Iberia]
        address:
          type: string
        floor:
          type: integer
        surface:
          type: number
        rooms:
          type: integer
        bathrooms:
          type: integer
        rentAmount:
          type: number
        chargesAmount:
          type: number
        depositAmount:
          type: number
        status:
          type: string
          enum: [disponible, occupe, maintenance, reserve]
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: string
        currentTenantId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - reference
        - title
        - type
        - zone
        - address
        - surface
        - rentAmount
        - status
        - createdAt
    CreatePropertyBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        zone:
          type: string
        address:
          type: string
        floor:
          type: integer
        surface:
          type: number
        rooms:
          type: integer
        bathrooms:
          type: integer
        rentAmount:
          type: number
        chargesAmount:
          type: number
        depositAmount:
          type: number
        status:
          type: string
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: string
      required:
        - title
        - type
        - zone
        - address
        - surface
        - rentAmount
        - status
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        cin:
          type: string
        email:
          type: string
        phone:
          type: string
        profession:
          type: string
        nationality:
          type: string
        dateOfBirth:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        status:
          type: string
          enum: [actif, inactif, en_retard, expulse]
        propertyId:
          type: integer
        propertyTitle:
          type: string
        rentAmount:
          type: number
        balance:
          type: number
        paymentScore:
          type: integer
        documents:
          type: array
          items:
            type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - cin
        - email
        - phone
        - status
        - createdAt
    CreateTenantBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        cin:
          type: string
        email:
          type: string
        phone:
          type: string
        profession:
          type: string
        nationality:
          type: string
        dateOfBirth:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
        status:
          type: string
        propertyId:
          type: integer
        notes:
          type: string
      required:
        - firstName
        - lastName
        - cin
        - email
        - phone
        - status
    Payment:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        tenantId:
          type: integer
        tenantName:
          type: string
        propertyId:
          type: integer
        propertyTitle:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: string
        status:
          type: string
          enum: [paye, en_retard, partiel, en_attente, annule]
        paymentMethod:
          type: string
          enum: [virement, cheque, especes, mobile_payment]
        penaltyAmount:
          type: number
        receiptNumber:
          type: string
        notes:
          type: string
        month:
          type: string
        createdAt:
          type: string
      required:
        - id
        - reference
        - tenantId
        - propertyId
        - amount
        - dueDate
        - status
        - month
        - createdAt
...